chore: Add [shared-type] path and file
